--- dashQC_fmri/viz/__init__.py
@@ -1,5 +1,0 @@
-"""For having the version."""
-
-import pkg_resources
-
-__version__ = pkg_resources.require("dashQC_fmri")[0].version
