--- dashQC_fmri/viz/config.py
@@ -1,11 +1,0 @@
-"""Default input parameters used in command line inface."""
-
-mode = 'normal'
-modes = ['normal', 'pseudocolor']
-fps = 20
-cmap = 'hot'
-sliceCor = 40
-sliceSag = 40
-sliceAx = 40
-slices = [sliceCor, sliceSag, sliceAx]
-slicesOrder = 'csa'
